One value per tile in multi-channel tessellation
To significantly facilitate the processing, tiles in each channel are
contiguous. So the tile map would only correspond to the actual input
dataset when there was only one channel or one dimension. The
`gal_tile_full_values_for_disp' has been added to do the job: you give it
an array (of any type) that is assumed to have one value for each tile. You
also give it the tessellation structure. This function will then sort the
values in the necessary way so that even when channels are present, the
tile values will be displayed properly.

This can later be used for interpolation (when interpolation is meant to be
done within channels or over the whole image for example). Also, it makes
it very easy to make a FITS file of the tile values for inspection (thanks
to the new `gal_tile_full_values_write'). To make this second job more
easily configurable the new `--oneelempertile' option has been added so if
the user only wants one pixel/element per tile (and not a huge image with
mostly constant values), they can do so easily with this option.

Other work that has been done in this commit:

 - `gal_blank_initialize' is a new function to initialize an array with
   blank values corresponding to its type.

 - Until now, `GAL_TILE_PO_OISET' would need the C types as arguments, but
   now, it calculates the types of the input and output to find the proper
   type internally. This greatly simplified the code, since the caller
   doesn't have to worry about the type any more.

 - `gal_tile_block_write_const_value' is a new function to write the tile
   values (one value per tile) into an allocated block of memory (with the
   same size as the tile's allocated block of memory). It can write values
   of any type that don't necessarily cover the whole image.

 - `gal_tile_block_check_tiles' now uses `gal_tile_block_write_const_value'
   to write the tile IDs into the allocated block.