NaN magnitude will only blank the profile's area
In MakeProfiles, to have a constant blank region, it was necessary to call
the `flat' or `circumference' profiles along with the `--mforflatpix'
option. In any other case, if the given magnitude of the profile was NaN
(blank), it would be divided by all the pixels covereing the profile's box
and so so the whole box would become blank.

NaN (blank) is a very unique value and is often necessary (when you want to
mask some region of the image). But the user can easily forget to call the
`--mforflatpix' option. When it is not intended (the NaN was a mistake),
there is going to be a blank region in the image any way and the user will
notice it, it doesn't matter if it is an ellipse or a box/rectangle. So a
simple check was added at the last step of finalizing the profile's
box/tile (in `oneprofile_make'). When the brightness of the profile is NaN,
then only the non-zero pixels in the profiles will be set to NaN and the
zero-valued pixels will just remain what they were (0.0).

Note that with this change, `--mforflatpix' is still necessary for non-NaN
constant values. This commit only simplifies the job for NaN magnitudes
which is a unique value.

This change was motivated because of my own desire to make a blank region
in an image (to implement interpolation), but forgetting to use the
`--mforflatpix' option. The blank box in the image was un-expected for me
(as a user), so I thought this correction would be helpful in the spirit of
the Principle of least astonishment.