Allow for floating point errors in Crop comparisons
When the input image is not aligned with the WCS or the pixel scales along
each dimension are different, in WCS mode, Crop will complain and abort
(because it currently uses these assumptions). Until now this check was
done exactly, so even a small difference of, for example 1e-10 (which is
clearly due to floating point errors) would cause an error. So with this
commit, very small relative differences (less than 1e-6) are acceptable and
Crop won't complain.

In this commit, short explanation that was also added to the `--polygon'
option of Crop so users can use SAO ds9 (a FITS viewer) for defining the
polygon instead of having to type the values manually.

Both these suggestions were kindly made by Guillaume Mahler, so his name
has been added to the `THANKS' file.