Permutations and neighbor finding algorithms
As part of finalizing the interpolation, it was necessary to rebuild some
infra-structure which has been done in this commit. The two main necessary
infrastructures were: Neighbor-finding (for each data-element/pixel) and
permutations. They are both described below:

Until now, the neighbor finding algorithm would need to actually check
every neighbor to see if it is outside the dataset or not. With the new
implementation, it is only necessary to check if the pixel is on the edge
of the image or not just once. The result is then stored as bits, so more
distant neighbors can be checked using them (instead of checking 8-byte
indexs). Like before, it was also defined as a macro, in the
`gnuastro/dimension.h' (new name for `gnuastro/multidim.h') header.

Permutations are necessary because tiles are defined contiguously over each
channel (not over the whole input dataset). So when it was necessary to do
operations over the whole image (for example interpolating when the
channels are to be ignored, or saving as a FITS file), the tile values
would not be displayed properly. Until now, we didn't use permutations, but
a very complicated and slow process of building a new array.

The main reason permutations were chosen was that after doing some
operations it was necessary to reverse the transformation above and put the
values back in tile-order. But in the old implementation, it was necessary
to write a completely new large and complicated function which was time
consuming and also buggy. So permutations were chosen to do the job,
because after defining one permutation, it is possible to apply itself or
its inverse with some very simple functions. The GNU Scientific Library
(GSL) had some great functions for the job. But unfortunately GSL works
based on subjective types like `int' and `long' not fixed width types. So
it was not easy to convert it to Gnuastro's types. Therefore, its functions
were re-implemented in a type-agnostic manner (using `memcpy') as part of
Gnuastro in a new `gnuastro/permutation.h' header.

As part of the process the following changes were also made:

 - `gal_data_bit_print_stream' is a new function to print a bit stream of
   any address in memory.

 - `gal_dimension_increment' is a new function to calculate the distance
   necessary to increment an index along all dimensions.

 - `gal_dimension_num_neighbors' is a new function to return the maximum
   number of neighbors to a data-element/pixel, given a specific dimension.

 - Thanks to `gal_dimension_increment', `gal_dimension_index_to_coord' has
   been re-implemented in a more robust/clean way.

 - `gal_tile_full_interpolate' is a new function to do interpolation over
   tiles. Since tile value interpolation is commonly necessary in a lot of
   places in Gnuastro and it is not too trivial, this function will do
   everything for any application.

 - Only some ground-work has started for the `lib/interpolate.c'
   library. But thanks to the major new infra-structure of this commit,
   work will go on much more smoothly after this commit.