Old neighbors.h header removed, used new in MakeProfiles
In the previous commit, a new macro was defined to easily parse over the
neighbors of an index in an array. But the old `lib/neighbors.h' header
still existed and was used in MakeProfiles to build the profile from the
center outwards. So as a good test for that macro and also a cleaner
implementation of MakeProfiles, the new macro is now used and the old
`lib/neighbors.h' header has been completely removed.