New inplementation of interpolation complete
An initial working implementation of interpolation with the new approach is
now complete. In the previous method, interpolation was heavily intertwined
with the tile structure, but now it can be done completely independent of
it. However, it is possible for the user to give a
`gal_tile_two_layer_params' structure to benefit from (for example to only
interpolate over a channel).

Other updates in this commit:

 - Also in this commit, the old `gal_tile_function_on_threads' function was
   changed to the more generic `gal_threads_spin_off'. In practice, the old
   function just used the number of tiles, but the new version takes in the
   number of actions as one argument, so it is usable in any context.

 - The new `gal_dimension_dist_manhattan' function will calculate the
   manhattan distance of two coordinates.