SubtractSky program removed, Statistics estimates Sky
Thanks to Arithmetic, the `Subtract' in the SubtractSky program was useless
for a long time, but it still remained. So after the old SubtractSky was
re-written to work with all the new infra-structure, it was decided that
ultimately, finding the Sky value (and its standard deviation) over an
image is a statistical operation. As a result, the estimation of the Sky
value has been added as another operation for the Statistics program. The
book has also been corrected to accommodate all these changes.

In the process of merging the new infra-structure into the sky finding
function, several other issues were addressed

 - The default tile size was changed to 50 from 40. This was done because
   50 is a more common number (people might think there is a special reason
   for choosing 40 and get confused). Also, it allows for more pixels, thus
   getting better estimates of the measurements on the tiles.

 - Tile interpolation options have been moved to the common arguments,
   since they are also needed in NoiseChisel.

 - The new `gal_options_read_sigma_clip' function will parse the values
   given to sigmaclipping parameters.

 - When the mode symmetricity is not good (mode is not accurate),
   `gal_statistics_mode' will return an array of all NaN values.

 - `gal_statistics_sigma_clip' now also uses
   `gal_statistics_no_blank_sorted'. Before it had its own implementation
   which didn't account for tiles.

 - Several new functions in `lib/tile.c'.

 - The main Gnuastro general configuration file `gnuastro.conf' was moved
   to the `bin' directory from the `lib' directory. The files in the `lib'
   directory are meant to be for the library. So this file was out of place
   there. It is now managed by the top level `Makefile.am'.