Internal headers moved to special directory
Until now, Gnuastro's internal headers were simply in the `lib/'
directory. So when you wanted to include `options.h', the include directive
would be `#include <options.h>'. Since Gnuastro's installed libraries are
in a `gnuastro' directory, the fact that the internal library headers are
simply mixed with the `.c' files of all the libraries can be confusing for
a new inquirer into the code.

To fix this issue and also to allow further symmetricity and clarity of the
source code, a new `lib/gnuastro-internal/' directory was created and all
internal headers are now put into there. So the include directive is now
`#include <gnuastro-internal/options.h>', it is longer, but it provides a
better level of documentation and clarity.

Work is also ongoing on the re-write of NoiseChisel, it has currently
reached the stage of filling holes that is necessary to define
pseudo-detections.