Correction of convolution over channel borders
When convolution is done in each channel separately, the boundaries of the
channels cause a discontinuity. This is ofcourse the desired effect. But
afterwards, when the effect of the channels has been subtracted, we want to
remove those edges in an already convolved image (this is what happens in
NoiseChisel). Since convolution is the most time consuming step of
NoiseChisel (or many other tools), is a significant over-kill to convolve
the whole image again just to remove the edges!

To fix this issue, the new `gal_convolve_spatial_correct_ch_edge' function
was defined. It will only convolve the pixels on the edges of channels. So
all the other pixels in the image will be left untouched and it is possible
to correct an already convolved image really fast and easy. Since the
implementation of convolution was effectively the same as before, only some
corrections were made to the main spatial convolution function (the new
`gal_convolve_spatial_general'). As a result the `gal_convolve_spatial' and
`gal_convolve_spatial_correct_ch_edge' are now just wrappers for that more
general function.

This was done as part of work on NoiseChisel, and in the process (to make
things more tidy in NoiseChisel), the Sky and its STD calculation and
subtractky functions were moved to a new `sky.c' source file.