NoiseChisel pseudo-detections identified on threads
Before implementing the new NoiseChisel (with `gal_data_t'), it would
identify pseudo-detections (fill holes and open the result) on multiple
threads and be much more efficient. But in the new implementation, it done
everything on the whole image on one thread which was slow. So with this
commit, it now applies the same strategy as before and will do the job on
multiple threads. This was partly done in preparation for the segmentation,
where a similar multi-threaded approach (when asking for a check image) is
required.

The tiles that the pseudo-detections are found on are larger tiles than the
standard ones for sky or threshold definition. This was also necessary for
segmentation (when trying to find the clump S/N value based on the
undetected regions). In the process, a bug was found in the tessellation
(which would appear when only one tile was necessary for a dimension), and
it has been fixed.