Clump S/N threshold is now calculated on Sky clumps
The new re-write of NoiseChisel now merges the clump S/N values found in
each tile and uses that to find the final S/N value.

The previous version would just return the S/N threshold and a histogram of
the S/N distribution if the user wanted. Also, the clump labels in the
check image started from one on each tile which would make it very hard to
check the exact derived S/N for a specific clump. Now, if the user asks for
a check image or a the clump S/N table, the labels in the check image will
be changed to a unique value for each clump and the output table that is
created also has the same ID so users can exactly check the derived S/N
values for each sky clump now.

Some other minor corrections in this commit:

  - A correction is also necessary for the previous commit: the default
    value for the `minbfrac' option is was changed to 0.7, not 0.8.

  - The `README' file was corrected to remove `SubtractSky' since it is no
    longer one of Gnuastro's programs.

  - The `txt' value was added to the `tableformat' option's `doc' in the
    output of `--help'. In other applications where the output filename is
    determined by the user, it was not necessary, but here, the filename
    will depend on the value to this option, so it was necessary to
    explicitly describe `txt' as one option.