NoiseChisel's clumps grown and objects defined
The new implementation of NoiseChisel will now fully grow the clumps,
determine the connectivity of the clumps and identify which clumps should
be merged into one object. It will finally expand the grown regions to
completely label the diffuse detected area.

The old function to find connected objects through an adjacency matrix
which was only usable within NoiseChisel has now been moved to a separate
library function: `gal_binary_connected_adjacency_matrix' (in
`gnuastro/binary.h'). All the old NoiseChisel functions for the steps above
have also been re-written in a much more cleaner way using the new tools,
but they are specific to NoiseChisel and weren't be taken to the libraries.