Header and file corrections after clean bootsrap and build
After the previous commit, I noticed that old library headers still existed
on my system so on my computer, the existence of such headers wouldn't
cause any problem but on others it would. As a result I manually deleted
all the installed Gnuastro libraries and headers from my system, completely
cleaned the version controlled directory (with `git clean -fxd') and
bootstrapped and re-built Gnuastro again. I found several additional cases
of old headers that no longer existed in the source code and remove them.

Since the `.in' files are automatically excluded from the `git status'
outputs, I hadn't noticed that the `lib/gnuastro-internal/config.h.in' file
was not version controlled. So this commit also includes this file in the
Git history (before it was in the `lib/' directory, but all internal
headers were later moved to the `lib/gnuastro-internal' directory and this
file was missed).