wcslib_version and restrict config macros
To verify if `wcslib_version' is present, previously we would only check
the existence of `GAL_CONFIG_HAVE_WCSLIB_VERSION'. But this was changed and
the macro now always exists, it just has a value of 0 or 1. So on Systems
where this function isn't present, there were unidentified linker
errors. This has been corrected.

Also, to help speed up the low-level processing of the `array' element of
`gal_data_t', we defined it as a restricted pointer. However, when the C89
compiler doesn't have an actual `restrict', this would cause compiler
problems during the checks (and later for library users). This problem
didn't exist in Gnuastro's programs because Gnulib has some good tests to
avoid such problems. So the result of Gnulib's configure time checks have
been used to define `restrict' in `gnulib/config.h'. This way, library
users can also benefit from the restrict pointer if their system supports
it. If it doesn't, `restrict' will be ignored (defined as an empty CPP
macro).

The dependency of the `mkcatalog/aperturephot.sh' test was missing a `.log'
so it would be skipped sometimes in parallel checks. This has been
corrected.

These issues were found and corrected with the great help of Alan Lefor on
some of his systems and his prompt remote checks. So I am very grateful to
him.

I am also working on updating the library documentation and some is
included in this commit.