Linkedlist library changed to list unified function names
Set of linked list functions were added over time and based on necessity,
so the weren't really too homogeous and for each type, the names were
different. Also the functions to treat `gal_data_t' as a linked list were
in the `data.h' header, not the linked list header. So with this commit,
all linked list operations have been moved to the new `list.h' header (and
`list.c' source file), and their names are now in the format of
`gal_list_TYPE_ACTION' which is much more easier to remember/manage than
the previous format of `gal_linkedlist_ACTION_CRYPTICTYPE' (since the type
was short and not very clear).

The old two-element double linked list has been removed with this
commit. It was only used in reading polygon vertices in Crop (in
`crop_polygonparser'), but now the same job is done more cleanly with an
ordinary `double' type singly linked list. The Linked list section of the
book/manual still needs some work, but the general context has been mostly
done.

Thanks to Guillaume Mahler, several warnings that showed up on Apple LLVM
version 8.1.0 (clang-802.0.42), mainly to do with printing (u)int64, were
also corrected in this commit. To solve it, we used the `inttypes.h'
header, so it is now also imported from Gnulib.

To parse over a tile, the `GAL_TILE_PARSE_OPERATE' works on all types. But
in some situations, the type is already known, so to avoid compiler errors
about, we are now directly using `GAL_TILE_PO_OISET'. This macro is now
corrected to be self-sufficient (not rely on variables in the calling
macro).