Completed Lists, types and blank library documentation
The documentation of the existing functions to work with lists has been
completed. In the process the `gal_list_tosizet_t' was renamed to
`gal_list_dosizet_t': instead of "two-way" we are now calling it
"doubly-linked", ordered size_t list which is the conventional name for
this type of lists. Some small corrections were also made to the functions
(that has now been applied in the manual).

Thanks again to Guillaume Mahler (Similar to the previous commit) another
compiler warning about printing `uint64_t' has been corrected.