Implementing __func__ in error messages
Work is continuing on documenting the library. In the process, I recognized
that Gnulib provides the `func' module that will enable us to use the
`__func__' pre-defined identifier. It is standardized in C99, but as Gnulib
discusses, it is not fully implemented on all systems, so it wasn't used in
Gnuastro. But after importing Gnulib's implementation of it, we can safely
use it even if the host system doesn't recognize it.

This new format of including the function name in the error messages has
now been implemented in nearly all library functions (except those that are
just about file opening or closing for example and can confuse the user
program users) and all internal function errors in programs: not errors
that directly relate to the user's inputs. For error messages that are
about the user's input, adding function names will most probably confuse
the user and reduce the affectiveness of the error message.