PC to CDELT decomposition done if pc present
Until now, the decomposition of PC (for rotation) and CDELT (for scale)
would only be done if WCSLB had marked the WCS structure as having a PCi_j
matrix through its `wcs->altlin' bitflag.

However, as noted by Alan Lefor, even when the input image only has CDi_j
matrix elements, WCSLIB will write out the (un-decomposed) PCi_j matrix in
the output. So apparently WCSLIB will have the pc matrix allocated and
filled even if the input is in CDi_j format. Hence with this commit, we
changed the condition for decomposig PC and CDELT matricese, from now on,
if `wcs->pc!=NULL', we will do the decomposition and change the
bit-flag. This worked on that example image.

In the process, I noticed that in MakeProfiles, we weren't setting the
`wcs->altlin' bitflag when making a WCS structure, this would cause issues
later (seen in `make check'). So it now sets this bit flag.