BuildProgram: to easily compile and link with Gnuastro's library
Gnuastro's library contains many wrappers to other libraries (for reading
FITS files for example) and managing all the linkings can be a big hurdle
to making small programs to check an idea or a hunch. So I created
BuildProgram within Gnuastro to greatly simply this job. It will use the
linking information that was gathered during the creation of Gnuastro to
build link with the user's program. In this way, the user doesn't have to
worry about linking.

BuildProgram also accepts the standard `-I', `-L', and `-l' compiler
options so if the user wants to use other libraries, it is easily
possible. This program is actually just a new implementation of the
`astcompilerun' shell function that was introduced in the old "Automatic
linking script" section of the manual (that section has now been
removed).

The main reason I chose to write a program for this job over a script is
Argp and all the option management features (for example with configuration
files) in Gnuastro's programs. It was not easy to implement all those
features with a shell script for me. So I just made a copy of the
`bin/TEMPLATE' directory and made this program instead. This also makes for
a similar user experience for all programs that start with `ast'.