Dimensions and Statistics libraries now documented
These two headers are now fully documented in the manual. An new library
demo was also added to show how to use `GAL_DIMENSION_NEIGHBOR_OP'. In the
process some argument names were also corrected (mainly for easier reading
and correspondance with the manual), for example `ind' was changed to
`index' and `data' was changed to `input'.