Threads and qsort documentation complete, demo for tables
The section discussing threads has also been moved to the second subsection
of all the libraries. This is done because it is indpendnet of a dataset
and only to do with processing. All the other libraries work on data.

The various `qsort' functions for all the types have been removed from the
manual and replaced with two generic functions `gal_qsort_TYPE_increasing'
and `gal_qsort_TYPE_decreasing'. This was done to make it readable, because
there were a large number of similarly named functions.

Two library demonstrations have been added to show how to read and write
columns in a table and how to write a multi-threaded program. They have
both been tested and hopefully show the most important relevent issues with
both kinds of operations.

Finally, the "GNU Free Documentation License" Appendix was renamed to "GNU
Free Doc. License" so it can fit witin the paddings of the table of
contents.