All libraries documented
All the libraries are now documented. In the process some other small
corrections were made in some remaining function APIs to make them more
consistent. Since their corrected version is now documented for the first
time with this commit, there is no need to go over them. But from this
commit onwards, any change in the library APIs must be documented in the
commit message (as well as the manual), so they can be included in the NEWS
file on every release.