Removed remaining linkedlist.h headers
Recently the `linkedlist.h' header was changed to `list.h'. But I had
forgot to remove some refences to it in the code. So if a previous version
of Gnuastro wasn't already available on the system, it would crash
(complaining about the non-existence of `linkedlist.h'). All such
occurances have been removed.

This was promptly reported by Alan Lefor and Guillaume Mahler after the
alpha-release annoucement.