Tests print cause of skip
Until now, when the necessary conditions (executable file or input images)
of a test were not present, it would just return a skip value. So it was
not clear why the test is being skipped in the `test-suite.log' file.

So with this commit, if any input condition is not present, before
returning a skip value and exiting, it is going to emit a message of what
is missing.

Thanks to a report by Boud Roukema, we found that the
`buildprog/simpleio.c' file that was an input to the
`buildprog/simpleio.sh' test was not being distributed in the tarball. This
has been corrected with this commit.