Width correction in gal_blank_as_string
After using the printf types of `inttypes.h', the `*' (to specify the width
of the print) was put in the `printf' formatting string in any case (when a
width was specified and when it wasn't). This is corrected with this
commit. This only affected the integer types.

The problem was actually found based on bad results from the checks in the
Table program when printing text tables. So on the path to finding the
cause (as described above), I also noticed that when setting the blank
string for each column, we were not using `gal_blank_as_string' for
strings! But now it supports strings also, so the extra
`gal_checkset_allocate_copy' has been removed and the code of this part is
much more cleaner now.

A small correction was also made in the comments of
`tests/buildprog/simpleio.c'.