Developing chapter of book edited/updated
The developing chapter was left almost intact during all the work that went
into the last ~200 commits. So I went through it and updated it to fit the
new system.

In the process, I also moved the two "Gnuastro project webpage" and
"Developing mailing lists" after the sections that described the
code. Previously, we would first discuss "Why C programming language" and
"Program design philosophy" (which are local issues). Then we would discuss
the project webpage and mailing lists, then return back to the code with
"Coding conventions" and the other sections to do with local things. Then
in the end we would go back to network issues with "Contributing to
Gnuastro". But with this change, the first few sections are fully devoted
to the developer's own work/computer and all the issues about networking
(with other Gnuastro developers) are moved to the end of the chapter. This
is much more logical.