Corrected ordering of table columns in Table
When the user wants a special set of the full list of columns, Table should
print them in the same order the user asked. However, since table was
implemented before the new option reading features, it would reverse the
list of input columns before read them. When the more recent version was
implemented, the option management tools in `option.c' do the reversing of
list input internally. So the programs don't need to worry about that.

So the final reversal (when columns were specified) was no longer needed
and would cause the final printed columns to be the opposite of the user's
request.