Corrections in --help output introductions
The first few paragraphs of the `--help' output of the programs was
reviewed and were necessary corrections were made. In particular, the
example commands to run in `fixedstingmacros.h' now have a `$' sign at the
start to more clearly show that they must be run on the command-line. The
rest of the corrections, were just program-specific edits/corrections.