Comments above example commands in Invoking sections
At the start of every "Invoking ProgramName" section in the manual, there
is a series of example commands to help the users see how to run the
program for different purposes. Until now, these were raw commands in all
programs except Table. However, these raw commands are hard to understand
for a new user (who isn't even sure what the program does!). So as
suggested by Rosa Calvi, a comment line was added above each command to
make them more easier to understand for a beginner and also better describe
the capabilities of the program in practice.

In the "Conventions" section of the manual/book, it is now also described
that a line beginning with `##' is a comment. While writing the comments,
several potential bugs were also found and corrected.

 - CosmicCalculator's `onlyvolume' option had a typo in its option
   definition and would have been called with `--onlyvolumne'!

 - When printing a final single number, Arithmetic would convert the final
   value's datatype into float32, but try printing the value as float64,
   hence printing terribly wrong values. It now converts to float64 and
   prints as float64.