Correct initialization of binary array to make profiles
During the building of a profile, MakeProfiles creates a binary array the
size of the desired profile and marks every pixel that has not been created
so it doesn't inspect one pixel more than once while looking at the
neighbors.

Previously we were allocating this array and clearning it manually, but
recently, it was replaced with the allocation functions within
`data.h'. However, instead of `gal_data_calloc_array', I had mistakenly
used `gal_data_malloc_array'. So the space for this array was not
initialized and would create strangely looking profiles (depending on what
was there in memory before). This bug was reported by Rosa Calvi and is
corrected with this commit.

Also, the comments in the Emacs `.dir-locals.el' file were all aligned for
easier reading.