-i option in Table takes precedence over -c
Previously, if a column was requested, the `-i' option would be
ignored. But it often happens that the users forget a column name after
already typing several of their desired columns. So the opposite behavior
is preferred. Because when more than a couple of columns are needed, you
will probably forget the column identifiers of the last few and having to
retype everything is very frustrating. Something like how `--help' takes
precedence over all other options.

Also, when printing a table to a text file, we were mistakenly using the
same precision for float32 and float64! This has been corrected now.