Function and variable name in gal_data_*alloc_array
If there was a problem in the allocation, the two `gal_data_malloc_arary'
and `gal_data_calloc_arary' functions would just abort with the size of the
necessary space. But that was not very useful: which function needed this
space? For which variable?

So with this commit, these two functions now have two extra arguments (that
may be NULL). These represent the function and variable names. When there
is an allocation error, these names will be used in the error message to
help in identifying exactly which step of a program confronted this issue.