Corrected comments in configuration files
The configuration file comments had not been updated since the updating of
the option management system. So with this commit, it was re-written and
copied/used for all the programs. The end of the top description of
`--help' is also updated with the suggested command descriptions listed on
the same line as the command.

In ConvertType, I noticed that `p->quality' (which accepts a value between
0 and 100) is being initialized with `UINT8_MAX' (because 0 is a valid
value for it). But this value is only meant as a blank value (to identify
if the option is actually given). So to be more clear, it is replaced with
`GAL_BLANK_UINT8'.