CFITSIO type code corrected, maximum option value length
In `gal_fits_key_write_filename', there was a mistaken reference to
CFITSIO's `TSTRING' type code rather than Gnuastro's
`GAL_TYPE_STRING'. This has been corrected.

In Crop, we now increment the input image counter `crp->numimg' in the
high-level `imgmodecrop' and `wcsmodecrop' functions, not in `onecrop'. The
reason for this is that the value is ultimately decided and used in these
high-level functions, so having it in `onecrop' didn't help in readability.

Finally, when printing option values and their short documentation, there
is now a maximum limit on the length of the value. If a value's length is
longer, that option's documentation will be one space character away from
the value (and not aligned). But atleast the other options will be
reasonable (as described above `GAL_OPTIONS_MAX_VALUE_LEN').