NoiseChisel's issue with high gthresh fixed
When a high `--gthresh' value was given, NoiseChisel would free the space
allocated to keep the diffuse fluxes. However, this space was necessary for
later filling of the region. So now, it is no longer freed, but the size is
set to zero. In this way, the adjacency matrix step is ignored, and we can
directly jump to the final filling of the object.