NoiseChisel's labeling with large gthresh corrected
With a large `--gthresh' the previous commit would ignore the object and
clump relabeling, but that was not the way to go, because we still need the
separate clump labels to be set to 1. So in this commit, when no diffuse
flux is given, we will manually set `clumptoobj' and `numobjects' and let
same old algorithms continue.

In the process, some other issues were fixed:

 - In MakeCatalog, we were mistakenly setting the `minmapsize' option to
   hidden (so it won't be read and a set to a value of 0). This has been
   corrected.

 - NoiseChisel's `--help' documentation for `tableformat' was corrected to
   be shorter to easily display in one line.

 - Some typos in the documentation for the discussion on the Adjacency
   matrix were corrected.

 - `gal_data_initialize' was mistakenly printing the value of `dsize[i]'
   with `%ld' instead of `%zu'. This remains from the time that `dsize' was
   in long type.