MakeCatalog output clump magnitudes corrected
When NoiseChisel is run with `grownclumps', detections that have one or no
clumps over them will be fully covered by a "clump". In that case, there
will be no rivers, so the magnitude and brightness measurements needed to
be corrected to account for no rivers.

Also, the output file names of MakeCatalog were corrected to deal with
directories in the value given to `--output'.

Finally, to aviod confusion, instead of having a negative S/N, MakeCatalog,
will just print a NaN value.