Corrected BuildProgram doc pointers in info and Man
The node name for BuildProgram's Invoking section was still unchanged from
its first version (`astcompilerun'). This would not allow `info
astbuildprog' to work properly. The node name is now corrected to `Invoking
astbuildprog'.

When Info couldn't find any menu item for `astbuildprog' it would look in
the Man-pages, but I noticed that it is bringing up the ConvertType Man
page instead! This was because of a mistake in creating of the Man-pages
which is now also corrected.