Arithmetic uses x instead of * for multiplication
Until now `*' represented multiplication in the Arithmetic
program. However, `*' is meaningful to the shell (all the files in a
directory), so the users would have to quote it when they wanted to
multiply inputs. So as suggested by Guillaume Mahler, we are now using the
easier to use `x' which is also more similar to the way we write
multiplication on paper. It is not really too likely someone will want to
work on a file named `x', so it is reasonable.

Other issues solved with this commit:

 - The main parameters structure of Arithmetic was still called
   `imgarithparams'! This is now corrected to `arithmeticparams'.

 - When reading the `ge' operator, we were mistakenly setting the operator
   to `GAL_ARITHMETIC_OP_LE'! This has been corrected.

 - For easier readability (mainly done in the process of finding the
   previous issue), the operators in `BINARY_RT_LT_SET' are now immediately
   under each other.