No blank lines in MakeCatalogs text table outputs
Until now, when the output was a text file, empty lines were present in the
middle of the comments. This is inconvenient for table readers like AWK,
you have to check for both comments and empty lines (and ignore
them). Without any empty lines, you can just ignore the commented ones. The
main problem was with the usage of `ctime' which prints a new-line
character in the end of its output. This new-line character is now manually
removed. Also the other comments don't add any new-line characters either.

Some other corrections in this commit:

 - The WCS structure is now read unconditionally in MakeCatalog. Until now,
   it would only be read when there was a WCS-related column. But it is
   also necessary to report the pixel area in the comments of the
   catalog. So that condition was removed.

 - When the user specifies a text file as output, the input value of
   `p->cp.tableformat' is now ignored and it is set to
   `GAL_TABLE_FORMAT_TXT'. This is done to help decide which comments
   should be added to the catalog in the end.

 - In `ui_check_options_and_arguments', `one' was not initialized! It is
   now set to `1' as it should have been. This caused a crash in one of our
   tests because the compiler had given it a value of zero.

 - In the option management system, the `set' flag should not be set for
   the `--config' option (which is unique). This has been corrected.